fishcoin/requirements.txt
cycler==0.10.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-filter==0.13.0
django-toolbelt==0.0.1
djangorestframework==3.3.3
gunicorn==19.4.5
httpie==0.9.3
Markdown==2.6.6
matplotlib==1.5.1
numpy==1.11.0
pandas==0.18.0
Pillow==3.2.0
psycopg2==2.6.1
Pygments==2.1.3
pyparsing==2.1.1
python-dateutil==2.5.3
pytz==2016.4
requests==2.9.1
scikit-learn==0.17.1
scipy==0.17.0
six==1.10.0
sklearn==0.0
static3==0.6.1
